requirements.txt
aniso8601==9.0.1
attrs==21.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
cryptography==35.0.0
Flask==2.0.2
flask-crontab==0.1.2
flask-restx==0.5.1
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==4.2.1
MarkupSafe==2.0.1
pycparser==2.21
PyJWT==2.3.0
pyrsistent==0.18.0
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
six==1.16.0
urllib3==1.26.7
uuid==1.30
websockets==10.1
Werkzeug==2.0.2
